init and start/finish request events on callbacks to instrumenter
