+++ lib/ex_debug_toolbar/plug/pipeline.ex
@@ -1,0 +1,6 @@
+defmodule ExDebugToolbar.Plug.Pipeline do
+  use Plug.Builder
+
+  plug ExDebugToolbar.Plug.ProcessRequestId
+  plug ExDebugToolbar.Plug.CodeInjector
+end
